requirements.txt
argon2-cffi
pycryptodome
python-networkmanager
